Move loaders to root of namespace
