+++ tests/Unit/DataSetLoaderTest.php
@@ -2,12 +2,12 @@
 /** @noinspection PhpUnhandledExceptionInspection */
 /** @noinspection PhpDocSignatureInspection */
 
-namespace webignition\BasilLoader\Tests\Unit\Loader;
+namespace webignition\BasilLoader\Tests\Unit;
 
+use webignition\BasilLoader\DataSetLoader;
 use webignition\BasilModel\DataSet\DataSet;
 use webignition\BasilModel\DataSet\DataSetCollection;
 use webignition\BasilModel\DataSet\DataSetCollectionInterface;
-use webignition\BasilLoader\Loader\DataSetLoader;
 use webignition\BasilLoader\Tests\Services\FixturePathFinder;
 
 class DataSetLoaderTest extends \PHPUnit\Framework\TestCase
