+++ tests/Unit/TestSuiteLoaderTest.php
@@ -2,7 +2,7 @@
 /** @noinspection PhpUnhandledExceptionInspection */
 /** @noinspection PhpDocSignatureInspection */
 
-namespace webignition\BasilLoader\Tests\Unit\Loader;
+namespace webignition\BasilLoader\Tests\Unit;
 
 use webignition\BasilDataStructure\PathResolver;
 use webignition\BasilModel\Assertion\Assertion;
@@ -17,7 +17,7 @@ use webignition\BasilModel\Value\ObjectNames;
 use webignition\BasilModel\Value\ObjectValue;
 use webignition\BasilModel\Value\ValueTypes;
 use webignition\BasilLoader\Exception\UnknownTestException;
-use webignition\BasilLoader\Loader\TestSuiteLoader;
+use webignition\BasilLoader\TestSuiteLoader;
 use webignition\BasilLoader\Tests\Services\FixturePathFinder;
 
 class TestSuiteLoaderTest extends \PHPUnit\Framework\TestCase
@@ -140,7 +140,7 @@ class TestSuiteLoaderTest extends \PHPUnit\Framework\TestCase
     {
         $expectedUnknownTestPath = (PathResolver::create())->resolve(
             __DIR__,
-            '../../../Fixtures/Test/example.com.path-does-not-exist.yml'
+            '../../Fixtures/Test/example.com.path-does-not-exist.yml'
         );
 
         $path = FixturePathFinder::find('TestSuite/example.com-path-does-not-exist.yml');
