+++ tests/Unit/YamlLoaderTest.php
@@ -2,12 +2,12 @@
 /** @noinspection PhpUnhandledExceptionInspection */
 /** @noinspection PhpDocSignatureInspection */
 
-namespace webignition\BasilLoader\Tests\Unit\Loader;
+namespace webignition\BasilLoader\Tests\Unit;
 
 use Symfony\Component\Yaml\Exception\ParseException;
 use Symfony\Component\Yaml\Parser as YamlParser;
 use webignition\BasilLoader\Exception\YamlLoaderException;
-use webignition\BasilLoader\Loader\YamlLoader;
+use webignition\BasilLoader\YamlLoader;
 use webignition\BasilLoader\Tests\Services\FixturePathFinder;
 
 class YamlLoaderTest extends \PHPUnit\Framework\TestCase
